requirements.txt
# IBKR API
ib_insync>=0.9.86

# Web Framework
dash>=2.14.0
dash-bootstrap-components>=1.5.0
dash-ag-grid>=31.0.0

# Charts & Data
plotly>=5.18.0
pandas>=2.0.0
numpy>=1.24.0

# Options Pricing
py_vollib>=1.0.1
scipy>=1.10.0
mibian>=0.1.3

# Database
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
